+++ src/windowMode.h
@@ -16,8 +16,8 @@ enum class processMode {
 class windowMode {
 private:
     processMode p_mode;
-    int last_min;
-    int last_max;
+    float last_min;
+    float last_max;
 
     std::vector<std::string> processNames = {
             "NORMAL",
@@ -32,8 +32,8 @@ public:
 
     void nextMode() { p_mode = static_cast<processMode>((static_cast<int>(p_mode) + 1) % 4); };
     std::vector<std::vector<float>> process(std::vector<StackSignal>& signals, std::set<int> selectedSignalsIndex, int last_n);
-    int get_last_min() { return last_min; };
-    int get_last_max() { return last_max; };
+    float get_last_min() { return last_min; };
+    float get_last_max() { return last_max; };
     std::string getModeName() { return processNames[static_cast<int>(p_mode)]; };
 
 };
